Add files via upload
Function to solve a linear equation of the form A.X=Y where A is a square matrix (N x N), Y is the y value vector and X is the X vector to solve for. A must be square and have no leading zeros within its rows. A different program is required for such matrices. Y and X must have the same column length, that is, N x 1 column vectors.